db/feeder/requirements.txt
aiomysql
faker
python-dotenv